+++ CHANGELOG.md
@@ -43,8 +43,8 @@
 - `react`                           *16.5.2*   →    **16.7.0**
 - `react-dom`                       *16.5.2*   →    **16.7.0**
 - `react-scripts`                    *1.1.5*   →     **2.1.3**
-- `react-swipeable-views`          *0.12.17*   →    **0.13.0**
-- `eslint-config-prettier`           *3.0.1*   →     **3.5.0**
+- `react-swipeable-views`          *0.12.17*   →    **0.13.1**
+- `eslint-config-prettier`           *3.0.1*   →     **3.6.0**
 - `eslint-plugin-prettier`           *2.6.2*   →     **3.0.1**
 
 ## [1.5.0] 2018-09-21